requirements.txt
flask>=2.2
openai-agents
duckduckgo-search
requests 